pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "followup-assistant"
version = "0.1.0"
description = "LinkedIn Follow-Up Assistant - Generate tailored follow-up messages"
authors = [{name = "User", email = "user@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer[all]>=0.9.0",
    "sqlalchemy>=2.0.0",
    "python-dotenv>=1.0.0",
    "rapidfuzz>=3.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
ai = [
    "sentence-transformers>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.20.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
followup = "src.app:app"
followup-web = "src.web:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"